+++ GildedRose/items/LegendaryItem.cs
@@ -1,7 +1,7 @@
 namespace GildedRose.Items{
     public class LegendaryItem : Item
     {
-        public LegendaryItem(string name, int sellIn, int quality, double basePrice) : base(name, sellIn, quality, basePrice)
+        public LegendaryItem(string name, int sellIn, int quality, int basePrice) : base(name, sellIn, quality, basePrice)
         {
 
         }
